changing double to int

+++ src/helloworld/Main.java
@@ -16,7 +16,7 @@ public class Main {
      */
     public static void main(String[] args) {
 
-double y=2;
+int y=2;
 
 System.out.println("Hello World! (Hussam!!!)");
